# impl: remove asserts causing classload