```
BEFORE:

Benchmark                                             Mode  Cnt     Score     Error        Units
ClassInitBenchmark.forName_init                         ss  400  8411.036 ± 148.305        us/op
ClassInitBenchmark.forName_init:·class.load             ss  400     1.136 ±   0.016  classes/sec
ClassInitBenchmark.forName_init:·class.load.norm        ss  400    71.040 ±   0.113   classes/op
ClassInitBenchmark.forName_init:·class.unload           ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_init:·class.unload.norm      ss  400       ≈ 0             classes/op
ClassInitBenchmark.forName_noinit                       ss  400  3077.689 ±  52.719        us/op
ClassInitBenchmark.forName_noinit:·class.load           ss  400     0.612 ±   0.007  classes/sec
ClassInitBenchmark.forName_noinit:·class.load.norm      ss  400    57.815 ±   0.093   classes/op
ClassInitBenchmark.forName_noinit:·class.unload         ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400       ≈ 0             classes/op

AFTER

Benchmark                                             Mode  Cnt     Score     Error        Units
ClassInitBenchmark.forName_init                         ss  400  7975.843 ± 114.276        us/op
ClassInitBenchmark.forName_init:·class.load             ss  400     1.080 ±   0.013  classes/sec
ClassInitBenchmark.forName_init:·class.load.norm        ss  400    69.763 ±   0.116   classes/op
ClassInitBenchmark.forName_init:·class.unload           ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_init:·class.unload.norm      ss  400       ≈ 0             classes/op
ClassInitBenchmark.forName_noinit                       ss  400  3103.897 ±  60.140        us/op
ClassInitBenchmark.forName_noinit:·class.load           ss  400     0.618 ±   0.009  classes/sec
ClassInitBenchmark.forName_noinit:·class.load.norm      ss  400    57.853 ±   0.086   classes/op
ClassInitBenchmark.forName_noinit:·class.unload         ss  400       ≈ 0            classes/sec
ClassInitBenchmark.forName_noinit:·class.unload.norm    ss  400       ≈ 0             classes/op

```